Add an INI file with texture load logging options
This will be useful for letting terrain mod developers determine which
textures are in use, and getting a log entry for missing textures.

The TerrainS3DTextureBindingFactoryProxy class that facilitates the
texture load logging can also be used to redirect the game's texture
loading based on the TGI of the texture.
This should also allow for a number of new features beyond the Maxis
terrain format.
Such as terrain mods including water or cutaway textures, expanding the
texture id range format, etc.