Add support for a DisplayNames section in terrain.ini
This section allows terrain mod developers to optionally provide a
display name for use tn the Extended Terrain UI.
The INI section entry key is the prefix name, and the value is the
display name.
Any terrain types that are not included will use the prefix name as the
display name.