Add a terrain texture id offset and redirect system
The texture redirect system allows some of the game's hard-coded
texture ids to be overridden in terrain.ini.
These new textures are specified using new fields in the terrain mod's
[<Prefix>MiscTextures] section. The supported textures are:

* 0x00000Z11 - CutawayTop
* 0x00000Z10 - CutawayMiddle
* 0x00000Z15 - CutawayBottom
* 0x00000Z14 - InshoreWater
* 0x00000Z13 - MidshoreWater
* 0x0918730Z - OutshoreWater

The texture offset system allows multiple terrain textures to use the
same 0x00000ZTT values in terrain.ini.
Because Maxis set the upper 20 bits of their terrain texture ids to 0,
we use those 20 bits as an offset to separate terrain textures that use
the same ZTT values.

The texture offsets are specified in a [TextureOffsets] section in
terrain.ini.
The field key is the texture prefix name, and the value is the offset.

A number of texture offsets that could conflict with the instance ids
of existing Maxis textures in the 0x7AB50E44, 0x891B0E1A group have
been marked as invalid, and will be rejected with an error message in
the log.